+++ tests/unit/test_main.py
@@ -9,15 +9,14 @@ class TestMain:
         mock_save_preferences = mocker.patch(
             "scanapi.settings.Settings.save_preferences"
         )
-        result = runner.invoke(main, ["--reporter", "markdown"])
+        result = runner.invoke(main, ["--output-path", "my_output.html"])
 
         assert result.exit_code == 0
         mock_save_preferences.assert_called_once_with(
             **{
                 "spec_path": None,
-                "output_path": None,
+                "output_path": "my_output.html",
                 "config_path": None,
-                "reporter": "markdown",
                 "template": None,
             }
         )