+++ tests/unit/test_reporter.py
@@ -18,18 +18,14 @@ class TestReporter:
         mocker.patch("scanapi.reporter.open", mock)
         return mock
 
-    @pytest.mark.parametrize("reporter_type", ["html", "markdown"])
     def test_should_write_to_default_output(
-        self, reporter_type, mocker, mocked__render_content, mocked_open
+        self, mocker, mocked__render_content, mocked_open
     ):
         mocked__render_content.return_value = "ScanAPI Report"
-        reporter = Reporter(None, reporter_type)
+        reporter = Reporter()
         reporter.write(fake_responses)
 
-        file_extension = {"html": "html", "markdown": "md"}[reporter_type]
-        mocked_open.assert_called_once_with(
-            f"scanapi-report.{file_extension}", "w", newline="\n"
-        )
+        mocked_open.assert_called_once_with("scanapi-report.html", "w", newline="\n")
         mocked_open().write.assert_called_once_with("ScanAPI Report")
 
     def test_should_write_to_custom_output(
@@ -48,8 +44,8 @@ class TestReporter:
         self, mocker, mocked__render_content, mocked_open
     ):
         mocked__render_content.return_value = "ScanAPI Report"
-        reporter = Reporter(None, None, "my-template.html")
+        reporter = Reporter(template="my-template.html")
         reporter.write(fake_responses)
 
-        mocked_open.assert_called_once_with("scanapi-report", "w", newline="\n")
+        mocked_open.assert_called_once_with("scanapi-report.html", "w", newline="\n")
         mocked_open().write.assert_called_once_with("ScanAPI Report")